配置到ｉｄｅａ：
