+++ src/bean/Img.java
@@ -1,0 +1,32 @@
+package bean;
+
+public class Img
+{
+    private String type;
+    private Appliance appliance;
+    private int id;
+
+    public String getType() {
+        return type;
+    }
+
+    public void setType(String type) {
+        this.type = type;
+    }
+
+    public Appliance getAppliance() {
+        return appliance;
+    }
+
+    public void setAppliance(Appliance appliance) {
+        this.appliance = appliance;
+    }
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+}
